admin: fix bug with confirmation during 1 week wait time
there were two different bugs that together made it impossible for
admins to confirm proposals made by other people before the waiting time
for allowing confirmations on self-made proposals had elapsed.

the one bug was that the confirm button was disabled in the interface.
the other was a database bug that was always entering the clause to exit
early for self proposals, regardless of whether the confirming admin was
the proposer or not