store sort and filter values in session cookies
revamped the solution for determining the sort order of threads on the
index page. instead of requiring the url params to always be present to
determine the order, we now use the latest setting and remember it in
the session cookies.

the reason behind this is so that we can combine sort order and filter
categories without having to go through too many gymnastics. the
resulting solution seems to be appropriate based on the small addition
it resulted in.