invites: remove allowlist and instead use invite paradigm for registration
this removes a lot of extra cruft from maaany places! one consequence of
staying consistent however is the creation of
"registration-instructions.md" and needing to update the local config in the
following way:

-verification_instructions = "content/verification-instructions.md"
+registration_instructions = "content/registration-instructions.md"

i.e. the 'verification_instructions' key is no longer, and is instead
reaplced with 'registration_instructions'. we could've kept the old
name, but in this instance considering the relatively few deployments
(known to me, at least) of cerca today. better to have a good name than
a legacy name in this instance

closes #39 #75 #47