invites: show tally of claimed invites by batchid
this is visible on /admin. for invites that have been claimed and whose
batch has either been completely exhausted or deleted by an admin, their
corresponding labels will be rendered as "unlabeled" which is why we
also make it possible to see the batch id.

we use "invite/register info" because for the first deployment of cerca,
urls were used as a way of noting incoming registrations. probably a
good thing to fix up the naming of in a future migration, but not
important enough to necessitate its own migration