Add invites functionality (#92)
* implement first pass of GetAllInvites routine
* remember to use references correctly in golang maps kids :):)
* use fmt.Errorf instead of errors.New(fmt.Sprintf) to write err descriptions
* invites: we crudding now - hook up database to server and html forms

we can now create and delete batches of invites, and render them in the
html view at /invites

i also decided to start contain server routes within constants written
in server/server.go, to make it easier to change them or whatever. just
feels less slightly messy than strings errywhere X)

* invites: remove allowlist and instead use invite paradigm for registration

this removes a lot of extra cruft from maaany places! one consequence of
staying consistent however is the creation of
"registration-instructions.md" and needing to update the local config in the
following way:

-verification_instructions = "content/verification-instructions.md"
+registration_instructions = "content/registration-instructions.md"

i.e. the 'verification_instructions' key is no longer, and is instead
reaplced with 'registration_instructions'. we could've kept the old
name, but in this instance considering the relatively few deployments
(known to me, at least) of cerca today. better to have a good name than
a legacy name in this instance

* invites: add descriptive text and log invite acts to modlog
* invites: implement reusable invites
* invites: show tally of claimed invites by batchid

this is visible on /admin. for invites that have been claimed and whose
batch has either been completely exhausted or deleted by an admin, their
corresponding labels will be rendered as "unlabeled" which is why we
also make it possible to see the batch id.

we use "invite/register info" because for the first deployment of cerca,
urls were used as a way of noting incoming registrations. probably a
good thing to fix up the naming of in a future migration, but not
important enough to necessitate its own migration

closes #39 #75 #47