db: refactor RemoveUser()
this refactor keeps the functionality of the previous RemoveUser
function intact while simultaneously structuring the code in a much
improved manner. the hope is to rewrite RemoveUser to take in booleans
make it more granular, such that posts can be kept intact during
deletion and/or usernames can be left intact as well while other
credentials are deleted.

for more context, see https://github.com/cblgh/cerca/issues/99#issue-2860668638