admin view: fix missing users in users list view
fixes #100

originally, at the start of the forum software, the only way for a user
to be created was through /register. over time the amount of ways a user
could be created grew, and in such a way that there could be no
registration info for them (such as when an admin manually creates a
user in /admin).

this caused a bug in the query for get users, where we would INNER JOIN
on the users table and the registrations table, and lose out on all the users
whom were lacking an entry in registration!

by changing that to a table users LEFT OUTER JOIN table registrations,
we'll get all of the entries from the left table (users) and supplement
with any intersections possible on the right side (registrations).