/account: wire up delete account options to the database
i've tested all permutations and the functionality seems to hold up!

[x] keep post, keep username
[x] keep post, remove username
[x] remove post, keep username
[x] remove post, remove username

one little word of warning on the keep username option: this account is
technically "retrievable" as in the user is still functional albeit
their password has been reset and not saved. however, it is possible
that e.g. an admin resets the user's password and can start using it.

in a future iteration, this can be blocked by adding a tombstone boolean
in the database. either by creating a new table (yay more tables) or by
augmenting the users table (this will necessitate a migration?)