add CERCA_ROOT and Config.Tooling.CercaRoot
from docs/hosting.md:

`CERCA_ROOT`

By default certain paths are relative to where `cerca` executes. These are:

* the css theme file that can be customized, `html/assets/theme.css`
* the content files defined by `config.Documents`, which are in `content/` by default
* the config file, called `cerca.toml` by default and can be defined with flag `--config`
* the database, residing in `data/forum.db` by default and an be defined with flag `--data`

In order to make these default assumptions more useful in different deployment environments,
the environment variable `CERCA_ROOT`. When set, it defines a base path from which to resolve
the above relative paths.

If any path (for example `Config.Document.About` is set to `/var/www/forum-content`) is an
absolute path, or if flags `--config` and `--data` are set to non-default values, then those
paths **will not** be joined with `CERCA_ROOT`.