use data_dir as source of truth for deployed docs/assets
Note: This commit also simplifies the contributed Makefile to leverage Cerca's
already existing behaviour for creating defaults. One primary reason for this
is that even ad-hoc usage of Cerca should be supported, so using the built-in
behaviour lets us not stray too far from that idea and (hopefully:) minimizes
distracting Makefile maintenance.

The changes in this commit were originally outlined in: https://github.com/cblgh/cerca/pull/117/files#r2187093620

That comment has been used as a description for this commit, with some light editing.

Make the `assets` directory use a path derived from `data_dir`:

// server/server.go
+	dbPath := filepath.Join(s.directory(), "forum.db")
+	docsPath := filepath.Join(s.directory(), "docs")
+	assetsPath := filepath.Join(s.directory(), "assets")
// ...
-    fileserver := http.FileServer(http.Dir("html/assets/"))
+    fileserver := http.FileServer(http.Dir(assetsPath))

As we can see above, we use `data_dir` as the source of truth also for document
paths (e.g. `Documents.LogoPath` etc). This lets us remove the `[documents]`
section of the config -- the reason `[documents]` was in the config in the
first place was to support different deployments on the same machine.

After this suite of changes we have the following logical structure:

```
${data_dir}/forum.db    -- the forum database
${data_dir}/docs/       -- rules.md, about.md, registration.md
${data_dir}/assets/     -- theme.css, logo.html, favicon.png, opensearch.xml, and any other other .png and media files used by the forum
```

Having this in place, we've now addressed the core concern of enabling multiple
deployments (whether via systemd or thru ad-hoc sysadmin scripts) in a
directory-agnostic way.