add command cerca write-defaults to save config and content files
this commit introduces the cerca write-defaults command, which can be
invoked as:

    cerca write-defaults --config /home/my-user/my-configs/cerca.toml --data-dir /var/lib/cerca

it will output the config file to the specified location, set the value
for key `auth_key` using a randomly generated auth key (same invocation
as genauthkey). the config's key `data_dir` will also be set. finally,
all the files to be put in the data_dir -- forum.db, docs/about.md etc,
assets/theme.css etc -- will also be written.

output is provided as part of command invocation as a way to expose the
forum host to what data files exist and can be customized.

changed:

as a result of a user-initiated path of creating a config and the
required data files, the behaviour of util.LoadFile no longer calls
util.CreateIfNotExist.